Convert index buffer to use uint32

+++ RenderingEngine/mesh.h
@@ -37,7 +37,7 @@ public:
 	AABB aabb;
 
 	int indexSize;
-	void createVertexAndIndexBuffer(ID3D11Device* device, const std::vector<Vertex> &vertices, const std::vector<uint16_t> &indices);
+	void createVertexAndIndexBuffer(ID3D11Device* device, const std::vector<Vertex> &vertices, const std::vector<uint32_t> &indices);
 };
 
 class CustomMesh : public Mesh {
